+++ src/extension.ts
@@ -17,6 +17,7 @@ import handleEditorReplaceVariable from './handler/variable-convert/editor-subme
 import { handleQuickPick as handleQuickPickVariable } from './handler/variable-convert/quick-pick-handler';
 import { commands as variableCommands } from './core/variable-convert/types/SupportVariableCaseType';
 import * as CyclicConversionVariable from './core/variable-convert/cyclic-conversion';
+import { showConvertCaseDetailDialog } from './core/variable-convert/show-convert-case-order-dialog';
 
 // Path Convert
 import handleEditorReplacePath from './handler/path-convert/editor-submenu-handler';
@@ -25,7 +26,7 @@ import { commands as pathCommands } from './core/path-convert/types/SupportPathF
 import * as CyclicConversionPath from './core/path-convert/cyclic-conversion';
 
 // Common
-import { createStatusBarItem, updateStatusBarItemVisable } from './handler/status-bar-handler';
+import { createStatusBarItem, updateStatusBarItemVisible } from './handler/status-bar-handler';
 import { EOL } from './types/EOLType';
 import { getUserConfigurations } from './utils/user-configuration';
 
@@ -77,7 +78,7 @@ export function activate(context: vscode.ExtensionContext) {
 		}
 
 		// 判断是否展示状态栏按钮
-		updateStatusBarItemVisable(selectTextLength);
+		updateStatusBarItemVisible(selectTextLength);
 
 		// 循环转换：记录当前选中内容，并且进行转换
 		let eol: EOL = textEditor.document.eol === vscode.EndOfLine.CRLF ? '\r\n' : '\n';
@@ -104,7 +105,7 @@ export function activate(context: vscode.ExtensionContext) {
 			onTextEditorSelectionChangeCallback(textEditor, selections);
 		} else { // 进入 else 的场景举例: 从[代码编辑器]切换到[设置页]
 			// 判断是否展示状态栏按钮
-			updateStatusBarItemVisable(selectTextLength);
+			updateStatusBarItemVisible(selectTextLength);
 		}
 	});
 
@@ -155,6 +156,10 @@ export function activate(context: vscode.ExtensionContext) {
 		CyclicConversionVariable.nextOne();
 	});
 	context.subscriptions.push(loopConvertCaseNextDisposable);
+
+	// 注册显示格式顺序信息的命令
+	let showConvertCaseDetailDialogDisposable = vscode.commands.registerCommand('variable-conversion.showConvertCaseDetailDialog', showConvertCaseDetailDialog);
+	context.subscriptions.push(showConvertCaseDetailDialogDisposable);
 
 
 	/**