+++ src/handler/path-convert/quick-pick-handler.ts
@@ -2,7 +2,7 @@ import * as vscode from 'vscode';
 import QuickPickItemEx from "../types/QuickPickItemExType";
 import { QuickPickSupportCaseItem, quickPickSupportCases } from '../../core/path-convert/types/SupportPathFormatType';
 import { TransformTextResult } from '../../types/TransformTextResultType';
-import { transformMutliSelectionText } from '../../utils/transform';
+import { transformMultiSelectionText } from '../../utils/transform';
 import { EOL } from '../../types/EOLType';
 import { pathConversion } from '../../core/path-convert/conversion';
 import { isStringArrayEqual } from '../../utils/utils';
@@ -12,18 +12,22 @@ const QuickPickLabelMaxLength = 60;
 
 interface RecommendItem {
     conversionText: Array<string>
-    transforTo: string[]
+    transformTo: string[]
     keyword: string[]
 }
 
 /**
  * 弹出的提示
  *
+ * @param textList
+ * @param eol
+ * @param enabledQuickPickSupportCases
+ * @returns
  * @since 2024-12-14
  */
 function generateOptionsBasedOnText(textList: string[], eol: EOL, enabledQuickPickSupportCases: Array<QuickPickSupportCaseItem>): Array<QuickPickItemEx> {
     // Cut text 切割文本
-    const resultsList: Array<TransformTextResult[]> = transformMutliSelectionText(textList);
+    const resultsList: Array<TransformTextResult[]> = transformMultiSelectionText(textList);
 
     const mergeResultList: Array<RecommendItem> = [];
     for (const quickPick of enabledQuickPickSupportCases) {
@@ -39,14 +43,14 @@ function generateOptionsBasedOnText(textList: string[], eol: EOL, enabledQuickPi
         if (recommendItem === undefined) {
             let item: RecommendItem = {
                 conversionText: conversionResults,
-                transforTo: [quickPick.shortName], // quickPick.name
+                transformTo: [quickPick.shortName], // quickPick.name
                 keyword: quickPick.keyword,
             };
             mergeResultList.push(item);
             continue;
         }
 
-        recommendItem.transforTo.push(quickPick.shortName); // quickPick.name
+        recommendItem.transformTo.push(quickPick.shortName); // quickPick.name
         recommendItem.keyword = Array.from(new Set(recommendItem.keyword.concat(quickPick.keyword))); // 关键词去重
     }
 
@@ -62,7 +66,7 @@ function generateOptionsBasedOnText(textList: string[], eol: EOL, enabledQuickPi
             label: conversionTextForDisplay.length >= QuickPickLabelMaxLength
                 ? (conversionTextForDisplay.substring(0, QuickPickLabelMaxLength - 3) + '...')
                 : conversionTextForDisplay,
-            description: `转换为 ${recommendItem.transforTo.join(' / ')}`,
+            description: `转换为 ${recommendItem.transformTo.join(' / ')}`,
             detail: `关键词 ${recommendItem.keyword.join(' ')}`,
             value: recommendItem.conversionText,
         };